Changing _check_db_availability function in run.sh
The old version of the _check_db_availability function is DB specific
and assuming that dsmr-reader uses postgresql. This is true for most of
the cases. 
Hovewer...
The dsmr-reader is build on top of django which supports different database
backends. This change is making _check_db_availability function universal and 
DB agnostic.