Turn nginx conf.d into real directory
Symlinked to http.d it causes configuration error. Nginx would read the
conf.d/, find dsmr-webinterface.conf and try to include it. Nginx would
find the upstream statement block in the wrong spot and error out.

Turning conf.d in a real directory instead of a symlink retains
functionality for those who intend to put files in there. And
dsmr-webinterface.conf is now only found within the http.d/ directory
and included in the correct context.